% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g_gradient}
\alias{g_gradient}
\title{Gradient edgelist}
\format{
\code{data.frame} edgelist
\itemize{
\item \code{from}
\item \code{to}
\item \code{weight}
}
}
\source{
Paper
}
\usage{
g_gradient
}
\description{
Gradient edgelist
}
\keyword{datasets}
