% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g_nongradient}
\alias{g_nongradient}
\title{Directed edge list with cycles}
\format{
An object of class \code{data.frame} with 13 rows and 3 columns.
}
\usage{
g_nongradient
}
\description{
Directed edge list with cycles
}
\keyword{datasets}
