% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hodgedecompose-package.R
\docType{package}
\name{hodgedecompose-package}
\alias{hodgedecompose}
\alias{hodgedecompose-package}
\title{hodgedecompose: Decompose a Flow Graph}
\description{
Decomposes a directed weighted graph into a directed acyclic flow, 
   a local cyclic flow and a global cyclic flow as described in <doi:10.1109/MCSE.2012.91>
}
\references{
https://facultyweb.cs.wwu.edu/~johnsoj5/prepub/Hodge\%20Tutorial.pdf
}
\author{
\strong{Maintainer}: Edwin de Jonge \email{edwindjonge@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
