% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nl_migration}
\alias{nl_migration}
\title{Internal Migration NL}
\format{
\code{data.frame} edgelist
\itemize{
\item \code{from}
\item \code{to}
\item \code{weight}
}
}
\source{
statline
}
\usage{
nl_migration
}
\description{
Internal Migration NL
}
\keyword{datasets}
