% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g_gradient}
\alias{g_gradient}
\title{Data set with only gradient flow}
\format{
An object of class \code{data.frame} with 13 rows and 3 columns.
}
\usage{
g_gradient
}
\description{
Data set with only gradient flow
}
\keyword{datasets}
